Separate mrc support from version field
We were overloading mrc_version = 0 to indicate
no support for MRC. It may be useful instead to use
mrc_version = 0 to let the provider choose the version
that should be used for the QP.

The MRC support is moved to a different output field in
mrc_query_device().

Accordingly, the protocol version has been added to QP attribute
in mrc_qp_attr, and associated flags.

Signed-off-by: Sayantan Sur <ssur@nvidia.com>